osd log added

+++ src/osd/OSD.h
@@ -1116,11 +1116,11 @@ public:
     static std::vector<bool> read_vec;
     static std::vector<bool> write_vec;
     static std::vector<bool> write_full_vec;
-    static std::vector<int64_t> op_dispatched_vec;
-    static std::vector<int64_t> op_enqueued_vec;
-    static std::vector<int64_t> op_dequeued_vec;
-    static std::vector<int64_t> op_started_vec;
-    static std::vector<int64_t> op_done_vec;
+    static std::vector<double> op_dispatched_vec;
+    static std::vector<double> op_enqueued_vec;
+    static std::vector<double> op_dequeued_vec;
+    static std::vector<double> op_started_vec;
+    static std::vector<double> op_done_vec;
 
 protected:
